[refactor] 필요없는 주석 및 import 제거
