+++ com.hobbing.gateway/src/main/java/com/hobbing/gateway/infrastructure/client/AuthClient.java
@@ -1,24 +1,19 @@
 package com.hobbing.gateway.infrastructure.client;
 
 import com.hobbing.gateway.domain.CustomHeader;
-import com.hobbing.gateway.domain.UserRole;
 import com.hobbing.gateway.dto.ApiResponse;
 import com.hobbing.gateway.dto.VerifyResponse;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
-import org.springframework.web.server.ResponseStatusException;
 import reactor.core.publisher.Mono;
 
 @Slf4j
 @Service
 public class AuthClient {
-//    @Value("${auth.host}")
+
     private String authHost = "http://localhost:19020";
 
     private final WebClient webClient;
@@ -34,8 +29,8 @@ public class AuthClient {
                 .header(CustomHeader.KEY_USER_ROLE, userRole)
                 .header(CustomHeader.KEY_INTERNAL_KEY, internalKey)
                 .retrieve()
-                .onStatus(HttpStatusCode::is4xxClientError, clientResponse -> {
-                    return Mono.error(new ResponseStatusException(HttpStatus.NOT_FOUND));})
+//                .onStatus(HttpStatusCode::is4xxClientError, clientResponse -> {
+//                    return Mono.error(new ResponseStatusException(HttpStatus.NOT_FOUND));})
                 .toEntity(new ParameterizedTypeReference<ApiResponse<VerifyResponse>>(){});
     }
 }
