+++ com.hobbing.gateway/src/main/java/com/hobbing/gateway/infrastructure/config/FeignConfig.java
@@ -7,8 +7,6 @@ import org.springframework.cloud.openfeign.support.SpringDecoder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-// FeignClient 를 Lazy Load 할 경우 HttpMessageConverters 문제로 아래 config 설정이 필요합니다.
-// open issue https://github.com/spring-cloud/spring-cloud-openfeign/issues/235
 @Configuration
 public class FeignConfig {
     @Bean
