+++ com.hobbing.gateway/src/main/java/com/hobbing/gateway/infrastructure/filter/AuthorizationFilter.java
@@ -1,0 +1,95 @@
+//package com.hobbing.gateway.infrastructure.filter;
+//
+//import com.hobbing.gateway.domain.CustomHeader;
+//import com.hobbing.gateway.domain.UrlEnum;
+//import com.hobbing.gateway.domain.UserRole;
+//import com.hobbing.gateway.dto.ApiResponse;
+//import com.hobbing.gateway.dto.VerifyResponse;
+//import com.hobbing.gateway.infrastructure.client.AuthClient;
+//import lombok.extern.slf4j.Slf4j;
+//import org.springframework.beans.factory.annotation.Value;
+//import org.springframework.cloud.gateway.filter.GatewayFilter;
+//import org.springframework.cloud.gateway.filter.factory.AbstractGatewayFilterFactory;
+//import org.springframework.http.HttpMethod;
+//import org.springframework.http.HttpStatus;
+//import org.springframework.http.server.PathContainer;
+//import org.springframework.http.server.reactive.ServerHttpRequest;
+//import org.springframework.stereotype.Component;
+//import org.springframework.web.server.ServerWebExchange;
+//import org.springframework.web.util.pattern.PathPattern;
+//import org.springframework.web.util.pattern.PathPatternParser;
+//import reactor.core.publisher.Mono;
+//
+//import java.util.Arrays;
+//
+//@Slf4j
+//@Component
+//public abstract class AuthorizationFilter extends AbstractGatewayFilterFactory<AuthorizationFilter.Config.class> {
+//
+//    @Value("${service.internal.internal-key}")
+//    private String INTERNAL_KEY;
+//
+//    private AuthClient authClient;
+//
+//    public AuthorizationFilter(AuthClient authClient) {
+//        this.authClient = authClient;
+//    }
+//
+////    @Override
+//    public GatewayFilter apply(AuthorizationFilter.Config config) {
+//        return (exchange, chain) -> {
+//            ServerHttpRequest request = exchange.getRequest();
+//            String path = request.getURI().getPath();
+//            HttpMethod method = request.getMethod();
+//            String userId = request.getHeaders().getFirst(CustomHeader.KEY_USER_ID);
+//            String userRole = request.getHeaders().getFirst(CustomHeader.KEY_USER_ROLE);
+//            String internalKey = request.getHeaders().getFirst(CustomHeader.KEY_INTERNAL_KEY);
+//
+//            if (userId == null || userRole == null || internalKey == null) {
+//                return errorResponse(exchange, "Missing required headers.");
+//            }
+//
+//            if (!internalKey.equals(INTERNAL_KEY)) {
+//                return errorResponse(exchange, "Invalid internal key.");
+//            }
+//
+//            if (!path.startsWith(UrlEnum.PREFIX_USERS.getUrl())) {
+//                return errorResponse(exchange, "Invalid path prefix.");
+//            }
+//
+//            ApiResponse<VerifyResponse> body = authClient.validateUserExists(userId, userRole, internalKey)
+//                    .block()
+//                    .getBody();
+//            VerifyResponse data  = body.data();
+//            if(data == null || !data.isVerified()){
+//                return errorResponse(exchange, "Permission denied.");
+//            }
+//
+//            if (checkPathPermissions(path, method, userRole)) {
+//                return chain.filter(exchange);
+//            } else {
+//                return errorResponse(exchange, "Permission denied.");
+//            }
+//        };
+//    }
+//
+//    public static class Config {}
+//
+//    protected abstract boolean checkPathPermissions(String path, HttpMethod method, String userRole);
+//
+//    protected boolean matchesPathPattern(PathPatternParser parser, PathContainer pathContainer, String pattern) {
+//        PathPattern pathPattern = parser.parse(pattern);
+//        return pathPattern.matches(pathContainer);
+//    }
+//
+//    protected boolean checkRole(UserRole[] userRoles, String role) {
+//        return Arrays.stream(userRoles).anyMatch(userRole -> userRole.name().equals(role));
+//    }
+//
+//    private Mono<Void> errorResponse(ServerWebExchange exchange, String message) {
+//        log.error(message);
+//        exchange.getResponse().setStatusCode(HttpStatus.FORBIDDEN);
+//        return exchange.getResponse().setComplete();
+//    }
+//
+//}
