+++ com.hobbing.gateway/src/main/java/com/hobbing/gateway/infrastructure/filter/UserAuthorizationFilter.java
@@ -6,16 +6,12 @@ import com.hobbing.gateway.domain.UserRole;
 import com.hobbing.gateway.dto.ApiResponse;
 import com.hobbing.gateway.dto.VerifyResponse;
 import com.hobbing.gateway.infrastructure.client.AuthClient;
-import com.hobbing.gateway.infrastructure.client.UserClient;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cloud.gateway.filter.GatewayFilter;
 import org.springframework.cloud.gateway.filter.factory.AbstractGatewayFilterFactory;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.http.server.PathContainer;
 import org.springframework.http.server.reactive.ServerHttpRequest;
 import org.springframework.stereotype.Component;
@@ -34,14 +30,8 @@ public class UserAuthorizationFilter
     @Value("${service.internal.internal-key}")
     private String INTERNAL_KEY;
 
-//    private UserClient userClient;
-
     private AuthClient authClient;
 
-//    public UserAuthorizationFilter(@Lazy UserClient userClient) {
-//        super(Config.class);
-//        this.userClient = userClient;
-//    }
     public UserAuthorizationFilter(AuthClient authClient) {
         super(Config.class);
         this.authClient = authClient;
@@ -57,7 +47,6 @@ public class UserAuthorizationFilter
             String userRole = request.getHeaders().getFirst(CustomHeader.KEY_USER_ROLE);
             String internalKey = request.getHeaders().getFirst(CustomHeader.KEY_INTERNAL_KEY);
 
-            // 에러 처리 메서드로 공통화
             if (userId == null || userRole == null || internalKey == null) {
                 return errorResponse(exchange, "Missing required headers.");
             }
@@ -70,8 +59,6 @@ public class UserAuthorizationFilter
                 return errorResponse(exchange, "Invalid path prefix.");
             }
 
-            // 권한 검증
-
             ApiResponse<VerifyResponse> body = authClient.validateUserExists(userId, userRole, internalKey)
                     .block()
                     .getBody();
@@ -79,26 +66,6 @@ public class UserAuthorizationFilter
             if(data == null || !data.isVerified()){
                 return errorResponse(exchange, "Permission denied.");
             }
-//            responseEntityMono.flatMap(resposeEntity -> {
-//                ApiResponse<VerifyResponse> apiResponse = resposeEntity.getBody();
-//
-//                VerifyResponse data = apiResponse.data();
-//
-//                data.isVerified();
-//                Mono.
-//
-//            });
-
-//            verifyResponseMono.
-
-//            ResponseEntity<ApiResponse<VerifyResponse>> verifyUser= userClient.verify(userId, UserRole.valueOf(userRole), internalKey);
-//            ResponseEntity<ApiResponse<VerifyResponse>> verifyUser= userClient.verify();
-//            VerifyResponse data = verifyUser.getBody().data();
-//            if(data == null || !data.isVerified()){
-//                return errorResponse(exchange, "Permission denied.");
-//            }
-
-
 
             if (checkPathPermissions(path, method, userRole)) {
                 return chain.filter(exchange);
@@ -114,7 +81,6 @@ public class UserAuthorizationFilter
         PathPatternParser patternParser = new PathPatternParser();
         PathContainer pathContainer = PathContainer.parsePath(path);
 
-        // Paths and roles validation
         if (matchesPathPattern(patternParser, pathContainer, "/users")) {
             return method == HttpMethod.GET && checkRole(new UserRole[]{UserRole.MASTER, UserRole.MANAGER, UserRole.TUTOR}, userRole);
         }
