+++ com.hobbing.proxygateway/src/main/java/com/hobbing/proxygateway/domain/CustomHeader.java
@@ -1,11 +1,8 @@
 package com.hobbing.proxygateway.domain;
 
 public class CustomHeader {
-    //    public final static String KEY_USER_ID = "X-User-Id";
     public final static String KEY_USER_ID = "user_id";
-    //    public final static String KEY_USER_ROLE = "X-User-Role";
     public final static String KEY_USER_ROLE = "user_role";
-    //    public final static String KEY_INTERNAL_KEY = "X-Internal-Key";
     public final static String KEY_INTERNAL_KEY = "secret_key";
     public final static String KEY_ACCESS_TOKEN = "Authorization";
     public final static String VALUE_BEARER_PREFIX = "Bearer ";
