+++ com.hobbing.user/src/main/java/com/hobbing/user/application/service/UserService.java
@@ -1,6 +1,5 @@
 package com.hobbing.user.application.service;
 
-import com.hobbing.user.application.dto.response.ApiResponse;
 import com.hobbing.user.application.dto.response.SearchedUsersResDto;
 import com.hobbing.user.application.dto.response.VerifyResponse;
 import com.hobbing.user.application.exception.UserErrorCode;
@@ -15,14 +14,8 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestMapping;
 
-import java.util.Optional;
 import java.util.UUID;
 
 @RequiredArgsConstructor
