--- com.hobbing.user/src/main/java/com/hobbing/user/infrastructure/filter/LogginFilter.java
@@ -1,4 +1,0 @@
-package com.hobbing.user.infrastructure.filter;
-
-public class LogginFilter {
-}
