+++ com.hobbing.user/src/main/java/com/hobbing/user/infrastructure/filter/LoggingFilter.java
@@ -1,0 +1,27 @@
+package com.hobbing.user.infrastructure.filter;
+
+import jakarta.servlet.*;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.core.annotation.Order;
+import org.springframework.stereotype.Component;
+
+import java.io.IOException;
+
+@Slf4j(topic = "LoggingFilter")
+@Component
+@Order(1)
+public class LoggingFilter implements Filter {
+    @Override
+    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
+        // 전처리
+        HttpServletRequest httpServletRequest = (HttpServletRequest) request;
+        String url = httpServletRequest.getRequestURI();
+        log.info(url);
+
+        chain.doFilter(request, response);
+
+        // 후처리
+        log.info("비즈니스 로직 완료");
+    }
+}
